refactor: api와 query 파일 위치 변경
