+++ src/hooks/query/useCommentAddQuery.ts
@@ -1,0 +1,22 @@
+import { useMutation, useQueryClient } from '@tanstack/react-query';
+import { CommentType } from 'src/store/types';
+import { addComment } from '@services/addComment';
+
+export interface AddCommentParams {
+  playlistId: string;
+  comment: CommentType;
+}
+// 댓글 추가 쿼리
+export const useCommentAddQuery = () => {
+  const queryClient = useQueryClient();
+
+  return useMutation({
+    mutationFn: (params: AddCommentParams) => addComment(params),
+    onSuccess: (playlistId) => {
+      queryClient.invalidateQueries({ queryKey: ['playlists', playlistId] });
+    },
+    onError: (error) => {
+      console.error('Failed to update error:', error);
+    },
+  });
+};
