+++ src/hooks/query/useCommentDeleteQuery.ts
@@ -1,16 +1,10 @@
-import { ref, remove } from 'firebase/database';
-import { db } from '@src/firebase';
+import { deleteComment } from '@services/deleteComment';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
-interface DeleteCommentParams {
+export interface DeleteCommentParams {
   playlistId: string;
   commentId: string;
 }
-const deleteComment = async ({ playlistId, commentId }: DeleteCommentParams) => {
-  const playlistRef = ref(db, `playlists/${playlistId}/comments/${commentId}`);
-  await remove(playlistRef);
-  return playlistId;
-};
 
 // 댓글 업데이트 쿼리
 export const useCommentDeleteQuery = () => {
