+++ src/hooks/query/usePlaylistAddQuery.ts
@@ -1,22 +1,7 @@
-import { ref, push, set } from 'firebase/database';
-import { db } from '@src/firebase';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { DefaultPlaylistType } from '@store/types';
-
 import { useUserStore } from '@store/useUserStore';
-import { fetchUserDataFirebase } from '@services/api/userService';
-
-const addPlaylist = async (newPlaylist: DefaultPlaylistType, userId: string) => {
-  const playlistRef = ref(db, `playlists`);
-  const newPlaylistRef = await push(playlistRef, newPlaylist);
-  const playlistId = newPlaylistRef.key;
-  const userRef = ref(db, `users/${userId}/createdPlaylists/${playlistId}`);
-  await set(userRef, true);
-
-  const updatedUserData = await fetchUserDataFirebase(userId);
-
-  return { playlistId, updatedUserData };
-};
+import { addPlaylist } from '@services/addPlaylist';
 
 // 플레이리스트 생성
 export const usePlaylistAddQuery = () => {
