+++ src/hooks/useSubscriptionAction.ts
@@ -1,4 +1,4 @@
-import { useSubscribedQuery } from '@services/reactQuery/useSubscribedQuery';
+import { useSubscribedQuery } from '@hooks/query/useSubscribedQuery';
 import { UserType } from '@store/types';
 
 export const useSubscriptionAction = (playlistId: string, user: UserType) => {
